SerialPortInfo, Windows implemented support get VID/PID
VID/PID in hexadecimal, returned as a string.
Reasons for the selected type of data as string:
1. Parameters VID/PID is initially stored and
retrieved from the OS as a string
2. There is a serial port (eg. not USB), which do
not have VID/PID in general. Therefore, they must
return an empty string. Other devices with VID/PID
of zero, should return a string as "0000". Therefore,
the use of VID/PID as a integer value (not as string)
is not correct.

Change-Id: I2b58d53d9d04673f36b09e4b3f6a5fabf06c69df
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>