Propagate vendor/productIdentifier terming instead of the vague vid/pid
* My opinion is based on personal biasement as the API user, plus things like
QRegularExpression recently added to Qt5.

* There are also other existing methods that have been using the long:
    http://doc-snapshot.qt-project.org/5.0/qnetworkconfiguration.html#identifier
    http://doc.qt.nokia.com/qt5/qsqldriver.html#IdentifierType-enum
    http://doc.qt.nokia.com/qt5/qsqldriver.html#escapeIdentifier
    ... and the like. There are many like those out there.

* Afair the coding style guide encourages to use full names instead of short
names, so I think this is reasonable

* Pid can be read at first glance like process identifier which is actually the
case inside the linux part of the codebase (that is used at few places)

Recap: All in all, it results a much nicer, better readable and memorizable API
for Qt developers. Tested on Linux and Windows, and worked fine. I am going to
ask my friend later for testing it on Mac whether there are any regressions. We
can fix those later, if any, but I do not expect any serious hazard which is not
straight-forward enough to fix, if any.

Change-Id: I08c0611a9953c5868ccf663df3294678d1545811
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>