Import the internal (private) QRingBuffer class from Qt5 to support Qt4
Making QRingBuffer public in Qt5 is not supported since it is not a real ring
buffer. It is used only for the QIODevice and QAbstractSocket internal buffers.
It is optimised for those. They will keep changing it to match those needs. That
is though not a real problem for our use case because the serial port module was
designed with QIODevice and QAbstractSocket resemblance in mind.

As for Qt4, even if it was possible, we cannot change this too flexibly, at
least not without a Qt4.9 and huge effort of convincing many people about it. I
have personally never heard of any plans about Qt4.9.

As for the time being, we could import this qringbuffer_p.h header file for
working out-of-the-box with Qt4. The file should be residing in the qt4support
folder. One neater solution would have been to use "core-private", but
unfortunately that is not gonna work with Qt4, as in:
"Project MESSAGE: Warning: unknown QT: core-private".

In any case, importing qringbuffer_p.h into the project seems to be a way as it
is used to be with QSerialDevice. This patch is yet another step against
superseding QSerialDevice with QtSerialPort even for Qt4 without too intrusive
changes in my opinion.

Change-Id: I95e1763f263ee4c94ba2634b097b94406564ac5b
Reviewed-by: Denis Shienkov <scapig@yandex.ru>