Import the internal QWinEventNotifier class from upstream Qt4 to support Qt4
QWinEventNotifier was made public since Qt5.0 for a good reason. See the
following links for details:
https://bugreports.qt-project.org/browse/QTBUG-68
https://bugreports.qt-project.org/browse/QTBUG-17218
http://doc.qt.nokia.com/qt5/qwineventnotifier.html

As for Qt4, even if it was possible, we cannot change this too flexibly, at
least not without a Qt4.9 and huge effort of convincing many people about it. I
have personally never heard of any plans about Qt4.9. Unfortunately my patch
had to be rejected because of the patch release BC/SC promises. We cannot
introduce new symbols and so forth. See my gerrit codereview request for
details about that:
http://codereview.qt-project.org/#change,22085

We could import this qwineventnotifier_p.h header file for working
out-of-the-box with Qt4. The file should be residing in the qt4support folder.
One neater solution would have been to use "core-private", but unfortunately
that is not gonna work with Qt4, as in:
"Project MESSAGE: Warning: unknown QT: core-private".

In any case, importing qwineventnotifier_p.h into the project seems to be a way
as it is also solved in the QtExtSerialPort project. This patch is yet another
step against superseding QSerialDevice with QtSerialPort even for Qt4 without
too intrusive changes in my opinion.

Change-Id: I40072a80b16ba8b83eca9f5eeb33c37866744ffb
Reviewed-by: Denis Shienkov <scapig@yandex.ru>