Use the established module name variable instead of hard coded strings
See the following documentation for details:
http://wiki.qt-project.org/Creating_a_new_module_or_tool_for_Qt#The_qt_.3Cmodule.3E.pri_files

Just one example out of there:
"QT.<module>.includes - Path of include/, typically $$QT_MODULE_INCLUDE_BASE
$$QT_MODULE_INCLUDE_BASE/$$QT.<module>.name"

Meanwhile, the module name is probably not going to change, it does not still
hurt to keep this in mind while developing. It is a minor issue though, but
makes the module a little bit more robust and professional.

Change-Id: Iabe16b52ffabcd46e3e19b27a17bfa4ced490946
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>