Add a helper script for installing the project header files with Qt4
Add a simple include helper .pri file for "generating" (copying) the public
header files into the projectroot/include folder. In addition, this helper
script takes care of the creation of that folder. Furthermore it sets up the
install target so that the "make install" execution will install the required
header files smoothly.

This is mostly done by the syncqt for the qt4 and qt5 codebases, but we cannot
expect people having any of those cloned. Moreover, it is a huge beast written
in perl. :-) Meaning that it is hard to adopt and make custom modifications,
aka. maintaining.

Change-Id: I19c0705c99600bfb8bb32d5755ddbe0b381be8d0
Reviewed-by: Denis Shienkov <scapig@yandex.ru>