Store the SERIALPORT_PROJECT_ROOT absolute path in the qmake cache
This is unfortunately a workaround for a bug addressed by issue in his patch
queue, but unfortunately just for Qt5. The basic problem is that $$OUT_PWD does
not return the build directory path. It will return the path where the $$OUT_PWD
is currently happening, and not inside where it would be expected.

In other words, I first experimented with LIBS += $$OUT_PWD/../ inside the
serialport.prf file for getting the library past in case an in-project build. It
always returned the absolute path location where the serialport.prf is actually
included, so every single example, demo, test and so forth. That is
unfortunately not sufficient since we need to store the absolute path to the
generated library more statically. The only workaround found for the time being
is storing this absolute path inside the qmake cache for the first run of qmake.

Change-Id: I0a0ec9616853d40bc0d54943dcf1227377642d5f
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>