Make the mkdir and copy operation cross-platform for qt4support
Use QMAKE_MKDIR and QMAKE_COPY instead of the raw "mkdir -p" and "cp -f"
commands. Those variables are properly turned into the platform specific bits.

Unfortunately there is a need for a separate replacement from the "\" directory
separator to the platform specific since the system commands otherwise might get
invalid syntax. This is the case with forward slashes in Windows environment for
example.

Change-Id: Ib58a455489a028e555183caaa3ee46789f723f82
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>