Show "Data bits: 8" in the settings of the terminal example by default
It is probably better to show this value in the settings for the data bits since
it is probably the most common usage out there. This way 8N1 would be available
by default. Therefore, in most cases while testing or debugging the qtserialport
codebase itself, it is not even needed to go to the settings, just press
connect.

Change-Id: I2cbe2e364ef488693c5687edf7658aa722e43a78
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>