Assign shortcuts to most of the actions available
It can be handy for certain people to use shortcuts for actions like in other
Desktop applications without too intrusive changes.

Selections:

* Connect: Ctrl+O -> Essentially because this shortcut is frequently used in
Desktop applications to open up a file. We do not open up files here, but the
serial port. Also, "o" is the second letter in the word "Connect". "C" might
be a poor choice since many words begin and contain that letter (Calls,
Connect, Configure and the like)

* Disconnect: Ctrl+D -> Essentially because this shortcut is frequently used in
terminal applications to close a session (ie. back from "root" to "user"). It is
also frequently used to close a process like an (for instance script language)
interpreter. In addition, this is the beginning letter of the "Disconnect" word.

* Configure: Alt+C -> Essentially because this shortcut is the most frequently
used for this action in other applications. We have actually made it default
inside our KDE Framework. It is pretty much the defacto standard for KDE
applications, and for many pure Qt applications as well. 'C' is also the
beginning letter of the word for this action.

* Clear: Alt+L -> Essentially because "Alt/Ctrl+C" would be confusing with
"Configure" and also with the generic "Copy" action. It is probably not the best
choice to accumulate the 'C' shortcut usage. "Ctrl+R" is widely used in
applications to refresh the window like browser. Hence, my decision is to use
"Alt+L" for this action. It is also a bit more difficult to type than "Alt+R"
meaning that it is less risky to clear the characters accidentally. This can be
reconsidered later, if the need arises. This is the hard selection out of the
all available. The other actions are straight-forward.

* About: Alt+A -> Essentially because this shortcut is the most frequently
used for this action in other applications. We have actually made it default
inside our KDE Framework. It is pretty much the defacto standard for KDE
applications, and for many pure Qt applications as well. 'A' is also the
beginning letter of the word for this action.

The relevant ampersand management is also being added so that the shorcut
character is underlined for the relevant menubar entry for the action in
question.

Change-Id: I64c206ff25f67cd5b7fb24528db8e1ae78e3d24b
Reviewed-by: Denis Shienkov <scapig@yandex.ru>