Use open/closeSerialPort terms instead of simply open/close
The reason behind is that, the close() method is already an existing one for the
QMainWindow. See the documentation here for further details:
http://qt-project.org/doc/qt-4.8/qwidget.html#close

I would like to use that method with a "Quit" action for closing the application
in the near future. See my next patch about the details of that feature.

Hence, renaming is a reasonable step in my opinion requiring minor effort. I
thought, if I am at renaming the implemented "close()" method to
"closeSerialPort()", I could just well rename the implemented "open()" method to
"openSerialPort()" as well for the sake of consistency even if there is no
"open" method for the QMainWindow class.

The relevant connect lines are also being modified obviously. I have tested the
connect and disconnect actions after this minor modification, and they work just
fine with my modem connected to the serial port via an usb-serial converter
(PL2303).

Change-Id: Ib5ee315e244cc5fa77210725b3548217ac12ca27
Reviewed-by: Denis Shienkov <scapig@yandex.ru>