Add a "Quit" action for closing the terminal example
Meanwhile, it is possible to close the application with window managers in
Desktop environment like KDE and so forth, it is sadly not that simple with
tiling window managers, like "Ratpoison". I have been using "Ratposion" for
almost 5-6 years by now. I have switched to this after KDE3 when KDE4 came out.

In such window managers, you do not have a "Top Bar" with the quit and other
icons, just a bare window about the content. Hence, I would like to propose
having a "quit" action for the application for make our lives simpler. My
workaround is sending a SIGKILL for such applications which is possible, but
surely not the nicest solution. :-)

The icon is adopted from the Oxygen Icons project that has also being used by
many KDE applications. In fact, this icon is pretty much the defacto standard
for the KDE applications. The icon size has just been chosen according to the
existing icon sizes inside the "$projectroot/examples/terminal/images" folder. I
am not sure we need size 128x128 in there, but I have put that size for
consistency with the rest. That probably does not hurt.

I do not have a rock solid vision about the placement of the action, but I have
currently put this into the first menu bar entry "Calls" with a separator, of
course, after the "Disconnect" action. I do not think we should establish a
separate menu bar entry just for this very action. I think it is reasonable this
way to have the "Quit" action in the first entry, until the need arises to
change this significantly in the near future that I doubt at this point.

Change-Id: I9f2639d4200e9a300deb872c8dbcf42d9dab54af
Reviewed-by: Denis Shienkov <scapig@yandex.ru>