Make the standardRates() method of the SerialPortInfo class static
The standard rates relate to the whole class in general instead of being
specific and tied to one instance. Hence, proposing the method to static.

In fact the Unix, Windows and Symbian serial port engines already use static
methods for this functionality, thus it is just an addition to the upper layer
in order to expose that as static to the SerialPortInfo API.

Change-Id: Ia3f8521e40a0858bfdadbd80121b4d49dfbed47e
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>