Fix the build of the examples on Linux
The root cause of the issue is that <QtAddOnSerialPort/*.h> cannot be found
unless the library is already installed. That is because the serialport.prf file
only provides the $$SERIALPORT_PROJECT_ROOT/include/QtAddOnSerialPort folder
for finding header files, but what the examples need according to the
aforementioned schema is "$$SERIALPORT_PROJECT_ROOT/include". Having made this
change, any examples, tests and demos can freely use either
<QtAddOnSerialPort/*.h> or <*.h> type includes in the future, so it remains to
be flexible this way.

Also, it is better to use "$$SERIALPORT_PROJECT_ROOT/include" instead of
"$$PWD/../../include" inside the serialport.prf file in order to have less
complexity and risk if we already have the serial port project root stored in a
relevant variable.

Tested on Linux and works fine even with QtCreator. The Windows debug and
release fix will come with the next change since that is a little bit separate
logic than the include directory fix.

Change-Id: I0f1ea7f5e0ae7f7580985fde53ab4405d4df2d89
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>