Use the qtAddLibary() and a different type of check for the serialport.prf
Let the builtin "qtAddLibary()" method manage the INCLUDEPATH and LIBS for this
module with Qt4 support so that we can make sure it is handled cross-platform
for Mac, Windows, Linux and Symbian. In additional, it has also some
consideration about the debug and release build type. The previous setup for the
installed module's INCLUDEPATH and LIBS are quirky and a bit hard coded ignoring
what the qtAddLibrary() qmake function already accomplishes.

Also, it is a bit cleaner logic to check against whether the module is already
installed, and if not prepend the "in-project" paths, as in the ones for
INCLUDEPATH and LIBS.

Tested on Linux and Windows. No regressions to the prior operations.
Unfortunately, it does not still provide proper release and debug builds, but
that comes with the next logical change. It is a dependency for that, and one
step ahead.

Change-Id: I6076d10e1d6c78605ea5eb5ac9fd918034b17fe8
Reviewed-by: Denis Shienkov <scapig@yandex.ru>