Fix the shadow build with Qt4
The reason is fairly simple: we need the top of the sources, and not the output.
The build of examples fail for out-of-tree builds. The issue to reproduce
without this patch is simple:

1) Go to the "$projectroot" folder.
2) mkdir -p build
3) cd build
4) qmake ../*.pro
5) make

Then there is yet another scenario to reproduce the issue, if the build
directory resides for instance in parallel with the $projectroot. The point is
that the build directory is outside the $projectroot, so not nested:

1) Go to the parent of the "$projectroot" folder.
2) mkdir -p qtserialport-build
3) cd qtserialport-build
4) qmake ../qtserialport/*.pro
5) make

After applying this patch and making sure the .qmake.cache file is deleted,
the build of the examples works. Tested on my Linux desktop box in command line,
and the patch works fine as expected.

Change-Id: I8713fc535bdb3dffff6fab64fe9ea3abb8d290ed
Reviewed-by: Denis Shienkov <scapig@yandex.ru>