Add license information to the header and source files
Having checked all the current Playground projects. This license selection seems
to be the most appropriate according to the other playground project licenses:
https://qt.gitorious.org/qtplayground/

In fact, QtSerialPort is the only Playground project which does not have license
yet. This patch tries to amend that gap. It is going to provide more trustworthy
situation about the usage of the module to commercial customers right away.
As for now, it looks a bit undefined what license this project goes for in the
end even if it is guessable. Commercial customers (perhaps even hobbyists) would
like to make sure they can reuse this library the way, they wish (aka. No GPLvX
and so forth).

Change-Id: Ia3a4e48a5140696f749df9fb5a8391955afaa637
Reviewed-by: Lars Knoll <lars.knoll@nokia.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>