On Windows, fix high CPU load while reading.
After success receiving notice of at least one character,
and his reading, then after this, for some reason always
occurred a false activation a notification subsystem,
even if the next new character to the port still did not
come. This is the cause of infinite activation a method
WinSerialPortEngine::event(QEvent *e) at each entry to the
event loop. This is also the reason for the increased load
on the CPU.

Most likely the reason for this behavior is the method
setReadNotificationEnabled(true), which is invoked from a
method SerialPort::readData(char * data, qint64 maxSize).

Change-Id: I5e709b7855784734f3d04d1eaf7735b5be3b823a
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>