Work around the build for unbreaking the linkage for the examples on Mac
Disclaimer: this is just a blind workaround, and not a real fancy technical
solution. There are people out there using Mac right away, like in our
commercial project. Working buildsystem for Mac is welcome.

The current issue that, the dylib library files are generated in the
$projectroot/src folder, but framework. This is an unfortunate situation since
the qtAddLibrary qt function deals with framework in a forced mode. See the
following link for details:
http://qt.gitorious.org/qt/qt/blobs/v4.8.2/mkspecs/features/qt_functions.prf#line25

There is also the following part which makes the Windows build work correctly,
but for the aforementioned forced reasons by default, this part of the qmake
function is not executed for Mac. Thereby, this part is copied out into our
qt4support/serialport.prf file.

Tests were done for the examples, like the terminal, including build and
runtime. Both worked with the same conditions as on Linux and Windows currently.

A prettier fix, or at least an alternative, could be that to figure out how to
build a framework with Qt4 and qmake functions for Mac. That question was asked
on the development@qt-project.org mailing list without getting any replies:
http://lists.qt-project.org/pipermail/development/2012-May/003813.html

Clearly, even a workaround is better for the commercial and pastime customers
than a non-working, half-baked and broken build system.

Change-Id: I997ae61b550d3463898d246d1ec033129fb44e14
Reviewed-by: Denis Shienkov <scapig@yandex.ru>