Establish a command line enumerator example mostly for limited environments
This is useful in environment where it is not possible or just not handy to run
UI examples to present the operation of the core SerialPort and SerialPortInfo
classes in the QtSerialPort add-on module. Typical use cases are embedded BSP
boards, like PandaBoard, BeagleBoard, Raspberry Pi and so forth.

Usually there is a "direct" connection over ssh or the native serial port to
those boards how the developer actually builds projects on the board. It would
be nice to build the QtSerialPort project in those cases, and then just fire a
command line example up.

In fact, while testing the QtSerialPort project on Mac, this is
also very useful at least to me since I have only limited access to a Mac box.
Perhaps, it can be the same with other people as well having access only to a
Linux box remotely and so forth.

Task-number: QTPLAYGROUND-1
Change-Id: If4681b05904845108ba2c6243b8955927211137c
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>