Add support for the gadget serial driver while using udev
The previous patch fixes the case when there is no available udev for managing
the list, but this also needs to be fixed. The problem is that, the devices are
disregarded without having the "usb" or "pnp" subsystem containment.

As for the /dev/ttyGS0 gadget serial driver, there is no subsystem, so that is
empty, but we still need to add those into the list as a valid serial port. The
only problematic case is when the subsystem is the "platform".

There are examples on my pandaboard for such invalid nodes that should not be
added, like the /dev/tty0* which are ARM OMAP related. After applying this
patch, the "cenumerator" example returns the gadget serial device properly on my
pandaboard with 12.04 stock ARM Ubuntu.

root@localhost:~/qtserialport# ./examples/cenumerator/cenumerator
Total number of ports available: 1

Port: ttyGS0
Location: /dev/ttyGS0
Description:
Manufacturer:
Vendor Idenifier:
Product Idenifier:
Busy: No
root@localhost:~/qtserialport#

Please see the following page for further information about the gadget serial driver:
http://kernel.org/doc/Documentation/usb/gadget_serial.txt

Change-Id: Ia7b17e1e0c294afec17658a1eb0971550dccc7b5
Reviewed-by: Denis Shienkov <scapig@yandex.ru>