Rename the include-helper to install-helper since it exceeds its original scope
This file was originally planned for helping with the include file generations,
but as the time went ahead, it got actually a bit more feature rich, and became
the centre of the helping with all the installation related features, like
mkspecs feature file, target library, dll binary and also of course the header
file installations.

Therefore, the current name is misleading while looking for information (i.e.
checking the target library installation related issues, or the prf file)

Change-Id: Ie714b858fbe47568c9a1e885e2cfb781f64f9cdf
Reviewed-by: Denis Shienkov <scapig@yandex.ru>