Build the UI examples only, if the QtWidgets module is present
This only works nicely with Qt5 because there is no simple way of achieving the
"!isEmpty(QT.widgets.name)" feature with the previous major versions.

In general, the recommendation is that to depend on modules in the examples that
are also the dependency of the library itself. However, that can be an exception
made with optional dependencies that the Qt Project accepts.

Tested on Linux with Qt5, and this works fine. There were no regressions to the
previous operations with Qt4 either.

Change-Id: I0d86557bb2a49300bb3bfd0adb94b61b1e7d4cc3
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>