The new concept of the internal structure
Complete replacement and optimization the internal logic,
because previous internal structure was complicated, not
flexible and sometimes causes more load of the CPU.

Changes:

* Replacement Engine extra layer of abstraction to a
private implementation. Now, all the platform-dependent
API (e.g. I/O, etc.) implemented in private classes, and
for I/O notification uses separate helper classes.

* Rejection of multiple inheritance of Notifiers classes.
Now for the notifications uses separate classes, which
simplifies the understanding of the code and to separate
these entities in terms of functionality. See, by analogy
to the implementation notifier in QAbstractSocket.

* The implementation of a fully asynchronous I/O for Win32.

* Only buffered mode supported

To try out the new concept can use a simple test-tool:
https://bugreports.qt-project.org/secure/attachment/28899/CpuUsageTester.zip

Change-Id: I372492ace3926e67671e7e0935312436d6ae8ac7
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>