Use an int maxSize api instead of a boolean byChunk
1) Simplifies the codebase since there is no any need for separate handlings
internally.

2) More flexible approach with regards to the size.

3) More consistent with the rest of Qt where maxSize is thoroughly used.

4) "byChunk" is a really weird term, but cannot be turn into anything marginly
better unfortunately.

5) The previously introduced way also has the term limitation that byChunk may
eventually mean that less than chunk size is written. The maxSize term is also
referring to that better.

6) This solution theoritically produces a faster code for time critical cases
where the chunk usage is preferred due to the avoiding of the extra condition
checks. It also procudes a bit slower performance for non time critical codes
like flush. In the end, this is what we need.

7) The compiler would not optimize this out either since it is beyond its scope.

8) No need for a brief documentation or comment for the method this way since
the parameter name makes sure it is clear what is taking place. With the other
approach, a short documentation or comment would have been a must have.

Change-Id: I60a9d3671f5c650f26c3add18d9bfc865a190c01
Reviewed-by: Denis Shienkov <scapig@yandex.ru>