Rejection of the reset() and replacing it an extended clear() method
The reset() method did not meet the description of the method of QIODevice,
he was incorrect implementation and functionality. Previously this method
do reset of all buffers, although he had to do it (see QIODevice doc):
"Seeks to the start of input for random-access devices. Returns true on
success; otherwise returns false (for example, if the device is not open)."

Therefore decided delete it. To replace the previous functionality and
extend it, have been introduced new method: "bool clear (Direction dir)".
Now the method have an input Direcrion parameter by which can clear the
desired buffers separately for different directions.

API of the operating system provides the possibility to clear the queue
separately in different directions, then it needs to be implemented now.
This may be necessary in the case of separate management of I/O data stream
for Tx / Rx. For example, in some protocols, it is necessary, also it
increases flexibility and versatility.

Change-Id: I0f63793189fcbbd80dfc8d9bc649c80c4eecc8da
Reviewed-by: Laszlo Papp <lpapp@kde.org>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>