Filter the symlinks for the /dev entries
There is no serialport device as symlink on Unix and hence it is safe to remove.
The explicit isDir() filtering can be eliminated this way.

It can be seen very well for the no udev case below that there are no serialport
entries with symlinks so the current handling is worse than it could be due to
the explicit filtering after the filtering method call.

    l << QLatin1String("ttyS*")    // Standart UART 8250 and etc.
      << QLatin1String("ttyUSB*")  // Usb/serial converters PL2303 and etc.
      << QLatin1String("ttyACM*")  // CDC_ACM converters (i.e. Mobile Phones).
      << QLatin1String("ttyGS*")   // Gadget serial device (i.e. Mobile Phones with gadget serial driver).
      << QLatin1String("ttyMI*")   // MOXA pci/serial converters.
      << QLatin1String("rfcomm*"); // Bluetooth serial device.
    l << QLatin1String("cu*");
    // Here for other *nix OS.

Change-Id: I8f23d9883c80cf6adff3890eef345215cd1133c3
Reviewed-by: Denis Shienkov <scapig@yandex.ru>