Getting info directly from the USB device
Previously, information about the product and the manufacturer of USB
device was received from the UDEV database. But UDEV database does not
always contain the correct information about the device. This
information may be outdated, and it may be absent in the case of "self-
made" USB devices.

A real example:

Many chips/microcontrollers (for example FTDI) can customize
settings of USB descriptor: VID, PID, iProduct, iManufacturer etc.

So the best solution is to get this information directly from the USB
descriptor, i.e. in this case, the information provided by the device.

PS: Thanks to Vladimir Kost for the catch.

Change-Id: I982c7a1fbd80df58273722de6178b923dae041b2
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Kost Vladimir
Reviewed-by: Denis Shienkov <scapig@yandex.ru>