Windows: fix write for Bluetooth serial ports
Bluetooth serial ports (at least from the Microsoft stack) do not
support the event EV_TXEMPTY. Therefore, for this reason, the writing
to the port was stopped after transmission of the first chunk of data
from the internal ring buffer of class.

Because further write the chunk of data to the port should be called
after triggered EV_TXEMPTY event.

The solution is to abandon the EV_TXEMPTY from design in library.

Now, for the launch to send of the next chunk of data to the port is
the event from write completion overlapped structure. Asynchronous
transmission is performed as long as the internal ring buffer is not
empty.

This solution also reduces the load on the CPU because the deleted
event processing for EV_TXEMPTY.

Task-number: QTPLAYGROUND-8
Change-Id: Ib4e21824ea783e1fe64ce0fa0be49ef0e69a99a6
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Aleksei Timofeyev <aleksei.timofeyev@gmail.com>
Reviewed-by: Denis Shienkov <scapig@yandex.ru>