Windows: Fix duplicate write data when call flush()
Previously, when using the flush() was called again startAsyncWrite() method,
which is extracted from the circular buffer incorrect data to be transmitted
(equal previous transferred data). I.e. reason for this behavior is a call
startAsyncWrite() more than once in a row, when completeAsyncWrite() do not
have time to process.

Now, every call startAsyncWrite() creates a separate WriteCompletionNotifier
with separate OVERLAPPED structure with control the position of bytes to write
in the ring buffer. This allows processing completionAsyncWrite() for each
separate chunk of transferred data.

Perhaps this will solve the problem:

Task-number: QTPLAYGROUND-10
Change-Id: If0590507a9cd0ab55ce53de82b4b1b3068833742
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>