Windows: Refactored notification subsystem
This is an important change when fixes critical bugs in the
algorithm of event processing when asynchronous or blocking I/O
and restructured notification classes.

Should be noted basic important changes:

* For all event notification classes created a common interface
AbstractOverlappedEventNotifier, which is the base for all the
necessary types of derived classes.

* All the necessary variables required for event notification
(e.g. OVERLAPPED structure and so forth) are moved into the
appropriate notifier classes.

* Of the common interface is now possible to know the object type
notifier (comm event, read or write) and also get a pointer to the
OVERLAPPED structure (that used for read/write operations).

* All objects of notifiers are now stored in the QHash, where the
key is a handle of event and value is pointer to notifier.

* To find the desired type of notifier from list now uses methods
such as lookupXXX().

* Creating a common interface notification allowed to use the same
data structures, for blocking and asynchronous I/O.

* Fixed blocking I/O when using methods like waitForXXX().

* Fixed a common algorithm for processing event notifiers when
reading data. Now the comm event notifier (which is used for the
detection of at least one Inbound bytes) runs only in the case of
an empty queue of driver. Until the driver queue is not empty,
works only read completion event notifier that tries to read all
the incoming data.

* For methods waitForReade() and waitForWrite() now uses the
common method waitAnyEvent(), which returns a pointer to triggered
notifier from list of notifiers.

Also this patch fix bug:

Task-number: QTPLAYGROUND-11
Change-Id: I9addd0b3f79877b76e858a9af17b13a9cf161a86
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>