Fix building all targets by default on Win32 and Mac for Qt4.x
Previously, for Qt4.x on Windows and Mac to build the project it was
necessary to separately specify the purpose Release and Debug, what
complicates the assembly, e.g. see Wiki:
http://qt-project.org/wiki/QtSerialPort#e200b03e6aa6d5363db5bdbb4386720c

Now, command "qmake qtserialport.pro" produces default Makefile,
that "(n)make && (n)make install" lead to the building and installation
of both Release and Debug targets for Windows and Mac.

Change-Id: I2b732aa7485eadf5a9082beb66b6024b7b5e6b1c
Reviewed-by: Laszlo Papp <lpapp@kde.org>
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>