Rename the SerialPort* classes to QSerialPort* including the internals.
Unfortunately, the rename has to happen in a separate commit because of the
limitations of "git":

"git does not track renames of files in history, git only records tree
snapshots. It can try to guess whether a change was a rename, or just unrelated
removal+addition, when asked. You can help this by making the rename in a
different commit from a modification. See man git-log -C and -M"

... so the rename will happen in the next change.

Change-Id: Ibe56eba0a0d2c7489db48a6c519d60d3315c00de
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>