Unbreak the terminal example compilation
This is a workaround, and not a real fix as the whole namespace situation is a
mess for now, and has to be discussed with others. The real fix has to wait, but
the build has to be working right now off-hand.

Also, there is not much point in exposing this global header file for Qt4 either
as the usage of the two public classes should be enough along with the
deterministic namespace name that will be agreed upon later (i.e. the namespace
define will be removed most likely).

Tested with Qt5, and the terminal example builds just fine on ArchLinux.

Change-Id: I9609afb43997ed20bf9770b517a93121bff84d82
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>