Fix the potential build breaks introduced by a previous commit (regression)
This is due to the fact the Qt namespace is not considered which is normally not
a problem as it is empty by default. However, using something like "-qtnamespace
MyNamespace" when configuring Qt will break the QtSerialPort build completely.

This is intended to be a fix for the library, and the examples/tests will be
fixed in a separate commit if necessary.

Change-Id: Iecd24dc5135c3a9645910754589dfd8bf4942337
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>